# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the Easy-Uni application.",
        "title": "Easy-Uni API",
        "contact": {}
    },
    "basePath": "/api",
    "paths": {
        "/events": {
            "get": {
                "description": "get all events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "parameters": [
                    {
                        "description": "Event",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateEvent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                }
            }
        },
        "/events/{event_id}": {
            "get": {
                "description": "get event by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/faculties": {
            "get": {
                "description": "get faculty by name, country, city, domain, budget",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faculties"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "name": "budget",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "domain",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Faculty"
                            }
                        }
                    }
                }
            }
        },
        "/faculties/{faculty_id}": {
            "get": {
                "description": "get faculty by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faculties"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Faculty ID",
                        "name": "faculty_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Faculty"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update faculty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faculties"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Faculty ID",
                        "name": "faculty_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Faculty",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateFaculty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Faculty"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete faculty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faculties"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Faculty ID",
                        "name": "faculty_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Faculty deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/faculties/{university_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "create faculty",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "faculties"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "University ID",
                        "name": "university_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Faculty",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateFaculty"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Faculty"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateEvent": {
            "type": "object",
            "required": [
                "name",
                "url",
                "visitor_id"
            ],
            "properties": {
                "campaign_id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "name": {
                    "type": "string"
                },
                "payload": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "url": {
                    "type": "string"
                },
                "visitor_id": {
                    "type": "string"
                }
            }
        },
        "models.CreateFaculty": {
            "type": "object",
            "required": [
                "budget",
                "domains",
                "name"
            ],
            "properties": {
                "about": {
                    "type": "string"
                },
                "academic_requirements": {
                    "type": "string"
                },
                "apply_date": {
                    "type": "string"
                },
                "budget": {
                    "type": "integer"
                },
                "domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "duration": {
                    "type": "number"
                },
                "language_requirements": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "other_requirements": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                }
            }
        },
...
        "models.Event": {
            "type": "object",
            "properties": {
                "campaign_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "name": {
                    "type": "string"
                },
                "payload": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "visitor_id": {
                    "type": "string"
                }
            }
        },
        "models.Faculty": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "academic_requirements": {
                    "type": "string"
                },
                "apply_date": {
                    "type": "string"
                },
                "budget": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "duration": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "language_requirements": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "other_requirements": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "university": {
                    "$ref": "#/definitions/models.University"
                },
                "university_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.JSONB": {
            "type": "object",
            "additionalProperties": true
        },
...
        "models.University": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "faculties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Faculty"
                    }
                },
                "id": {
                    "type": "string"
                },
                "img_link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ranking": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}